# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ticket-statuses": {
            "get": {
                "description": "Get a list of all ticket statuses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-statuses"
                ],
                "summary": "Get all ticket statuses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.TicketStatus"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-statuses"
                ],
                "summary": "Create a new ticket status",
                "parameters": [
                    {
                        "description": "Ticket Status Data",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TicketStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TicketStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ticket-statuses/{id}": {
            "get": {
                "description": "Get a ticket status by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-statuses"
                ],
                "summary": "Get a ticket status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TicketStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ticket status by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-statuses"
                ],
                "summary": "Update a ticket status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Status Data",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TicketStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TicketStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket status by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-statuses"
                ],
                "summary": "Delete a ticket status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        },
        "/tickets": {
            "get": {
                "description": "Get a list of all tickets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get all tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/requests.TicketResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Create a new ticket",
                "parameters": [
                    {
                        "description": "Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Get a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get a ticket by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ticket by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Update a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Delete a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "validation": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.AdminAvailability": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/models.User"
                },
                "admin_id": {
                    "type": "integer"
                },
                "current_conversations": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.AdminConversationState": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/models.User"
                },
                "admin_id": {
                    "type": "integer"
                },
                "conversation": {
                    "$ref": "#/definitions/models.Conversation"
                },
                "conversation_id": {
                    "type": "integer"
                },
                "last_message_id": {
                    "type": "integer"
                },
                "unread_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Conversation": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/models.User"
                },
                "admin_id": {
                    "type": "integer"
                },
                "admin_state": {
                    "$ref": "#/definitions/models.AdminConversationState"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/models.User"
                },
                "customer_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_message_at": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.ConversationStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ConversationStatus": {
            "type": "string",
            "enum": [
                "open",
                "closed"
            ],
            "x-enum-varnames": [
                "StatusOpen",
                "StatusClosed"
            ]
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "assignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAssignment"
                    }
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAttachment"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.TicketCategory"
                },
                "category_id": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketComment"
                    }
                },
                "deskripsi": {
                    "type": "string"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "kode_tiket": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketLog"
                    }
                },
                "priority": {
                    "$ref": "#/definitions/models.TicketPriority"
                },
                "priority_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.TicketStatus"
                },
                "status_id": {
                    "type": "integer"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "tanggal_diperbarui": {
                    "type": "string"
                },
                "tipe_pengaduan": {
                    "type": "string"
                },
                "user": {
                    "description": "Relasi - Add references to match custom column names",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.TicketAssignment": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/models.User"
                },
                "id_admin": {
                    "type": "integer"
                },
                "id_assignment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "tanggal_ditugaskan": {
                    "type": "string"
                },
                "ticket": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                }
            }
        },
        "models.TicketAttachment": {
            "type": "object",
            "properties": {
                "file_path": {
                    "type": "string"
                },
                "id_attachment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "ticket": {
                    "description": "Relasi - specify references",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                },
                "uploaded_at": {
                    "type": "string"
                }
            }
        },
        "models.TicketCategory": {
            "type": "object",
            "properties": {
                "id_category": {
                    "type": "integer"
                },
                "nama_category": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                }
            }
        },
        "models.TicketComment": {
            "type": "object",
            "properties": {
                "id_comment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "isi_pesan": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "ticket": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.TicketLog": {
            "type": "object",
            "properties": {
                "aktivitas": {
                    "type": "string"
                },
                "id_log": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "ticket": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
        "models.TicketPriority": {
            "type": "object",
            "properties": {
                "id_priority": {
                    "type": "integer"
                },
                "nama_priority": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                }
            }
        },
        "models.TicketStatus": {
            "type": "object",
            "properties": {
                "id_status": {
                    "type": "integer"
                },
                "nama_status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "admin_availability": {
                    "$ref": "#/definitions/models.AdminAvailability"
                },
                "admin_states": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AdminConversationState"
                    }
                },
                "assigned_tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAssignment"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketComment"
                    }
                },
                "conversations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Conversation"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketLog"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "seller",
                "customer"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleSeller",
                "RoleCustomer"
            ]
        },
...
        "requests.TicketCreateRequest": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string",
                    "example": "Saya tidak bisa login ke akun saya."
                },
                "id_category": {
                    "type": "integer",
                    "example": 1
                },
                "id_priority": {
                    "type": "integer",
                    "example": 1
                },
                "id_status": {
                    "type": "integer",
                    "example": 1
                },
                "id_user": {
                    "type": "integer",
                    "example": 1
                },
                "judul": {
                    "type": "string",
                    "example": "Tidak bisa login"
                },
                "kode_tiket": {
                    "type": "string",
                    "example": "TCKT-001"
                },
                "tipe_pengaduan": {
                    "description": "Allowed: pelanggan, penjual",
                    "type": "string",
                    "enum": [
                        "pelanggan",
                        "penjual"
                    ],
                    "example": "pelanggan"
                }
            }
        },
        "requests.TicketResponse": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "id_category": {
                    "type": "integer"
                },
                "id_priority": {
                    "type": "integer"
                },
                "id_status": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "kode_tiket": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "tanggal_diperbarui": {
                    "type": "string"
                },
                "tipe_pengaduan": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}